requirements.txt
fastapi[all]
uvicorn
firebase-admin
pydantic
starlette
python-dotenv
pyjwt
chainlit
